// Client/src/app/datenservice/AppDTO.ts
/* spiritdev Softwareentwicklung GmbH 2020
      Author Vladimir Kogan
*/
export interface AppDTO {
    id: number;
    title: string;
    annotation: string;
    price: number;
    image: string;
    category: string;
    plattform: string;
    projectStart: number;
    projectEnd: number;
    salesAmount: number;

}